requirements.txt
Django>=4.2
psycopg2-binary>=2.9
boto3>=1.26
django-storages>=1.13
djangorestframework==3.16.1
minio==7.2.7
django-filter>=23.5